requirements.txt
azure-functions==1.11.3
